TradingBot/Strategist/requirements.txt
loguru==0.6.0
pyzmq==25.0.0
numpy
requests